version2/requirements.txt
altair==4.2.2
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
astor==0.8.1
asttokens==2.2.1
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
base58==2.1.1
bcrypt==4.0.1
beautifulsoup4==4.12.1
bidict==0.22.1
black==23.3.0
bleach==6.0.0
blinker==1.6
boltons==23.0.0
bottle==0.12.25
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
contextlib2==21.6.0
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Django==4.2
django-admin==2.0.2
django-excel-response2==3.0.5
django-six==1.0.5
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
entrypoints==0.4
excel-base==1.0.4
executing==1.2.0
face==22.0.0
fastapi==0.92.0
fastapi-socketio==0.0.10
fastjsonschema==2.16.3
fonttools==4.39.3
fqdn==1.5.1
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==6.2.0
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
isoweek==1.3.3
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
loguru==0.6.0
markdown-it-py==2.2.0
markdown2==2.4.8
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mypy-extensions==1.0.0
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.0
nbformat==5.8.0
nest-asyncio==1.5.6
nicegui==1.2.6
notebook==6.5.4
notebook_shim==0.2.2
numpy==1.24.2
orjson==3.8.9
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
plotly==5.14.1
pocket-protector==20.0.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
proxy-tools==0.1.0
pscript==0.7.7
psutil==5.9.4
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg-pool==3.1.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
PyNaCl==1.5.0
pyobjc-core==9.0.1
pyobjc-framework-Cocoa==9.0.1
pyobjc-framework-WebKit==9.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.4.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
python-socketio==5.8.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywebview==4.0.2
PyYAML==6.0
pyzmq==25.0.2
regex==2023.3.23
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.3
rsa==4.9
ruamel.yaml==0.17.21
schema==0.7.5
scikit-learn==1.2.2
scipy==1.10.1
screen==1.0.1
seaborn==0.12.2
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.9
sqlparse==0.4.3
stack-data==0.6.2
starlette==0.25.0
streamlit==1.21.0
tenacity==8.2.2
terminado==0.17.1
testpath==0.6.0
threadpoolctl==3.1.0
TimeConvert==3.0.13
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uri-template==1.2.0
urllib3==1.26.15
uvicorn==0.20.0
uvloop==0.17.0
validators==0.20.0
vbuild==0.8.1
watchdog==3.0.0
watchfiles==0.18.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.1
widgetsnbextension==4.0.7
xlwt==1.3.0
zipp==3.15.0
